+++ lib/pbio/platform/essential_hub/pbsysconfig.h
@@ -15,6 +15,7 @@
 #define PBSYS_CONFIG_HUB_LIGHT_MATRIX               (0)
 #define PBSYS_CONFIG_HOST                           (1)
 #define PBSYS_CONFIG_HOST_EVENT_OUT_SIZE            (512)
+#define PBSYS_CONFIG_HUB_TYPE_STR                   "SPIKE Essential"
 #define PBSYS_CONFIG_MAIN                           (1)
 #define PBSYS_CONFIG_STORAGE                        (1)
 #define PBSYS_CONFIG_STORAGE_NUM_SLOTS              (1)