+++ lib/pbio/platform/nxt/pbsysconfig.h
@@ -14,6 +14,7 @@
 #define PBSYS_CONFIG_HMI_NUM_SLOTS                  (0)
 #define PBSYS_CONFIG_HOST                           (1)
 #define PBSYS_CONFIG_HOST_EVENT_OUT_SIZE            (512)
+#define PBSYS_CONFIG_HUB_TYPE_STR                   "MINDSTORMS NXT"
 #define PBSYS_CONFIG_MAIN                           (1)
 #define PBSYS_CONFIG_STORAGE                        (1)
 #define PBSYS_CONFIG_STORAGE_NUM_SLOTS              (1)