+++ lib/pbio/platform/virtual_hub/pbsysconfig.h
@@ -12,6 +12,7 @@
 #define PBSYS_CONFIG_BATTERY_CHARGER                (0)
 #define PBSYS_CONFIG_HOST                           (1)
 #define PBSYS_CONFIG_HOST_EVENT_OUT_SIZE            (512)
+#define PBSYS_CONFIG_HUB_TYPE_STR                   "Virtual Hub"
 #define PBSYS_CONFIG_HMI                            (1)
 #define PBSYS_CONFIG_HMI_STOP_BUTTON                (1 << 7) // center
 #define PBSYS_CONFIG_HMI_VIRTUAL                    (1)